+++ Services/CoreLibrary.cs
@@ -35,6 +35,7 @@ public class CoreLibrary
     public FileWatcherService FileWatcher { get; }
     public Action<string, object?> SendToJS { get; }
     public Action<string>? AvtrdbSubmit { get; set; }
+    public Action<string>? VrcndbSubmit { get; set; }
 
     public ConcurrentDictionary<string, bool> PlayerAgeVerifiedCache { get; } = new();
     public ConcurrentDictionary<string, Newtonsoft.Json.Linq.JObject> PlayerProfileCache { get; } = new();